\name{CentromeresHg19}
\docType{data}
\alias{CentromeresHg19}
\title{Positions of centromeres on all chromosomes, human genome 19}
\description{
CentromeresHg19 is a data frame providing positions of centromeres on all chromosomes of human genome 19 and its sizes. Size of centromere is also approximated in WinSize column to allow proper visualization of particular chromosomes.

}
\usage{data(CentromeresHg19)}
\value{test data}
\format{data frame}
\source{UCSC table browser}
\keyword{datasets}

