\name{SonRareVariantsChr21}
\docType{data}
\alias{SonRareVariantsChr21}
\title{Annotated rare variants on the chromosome 21, human genome 19, subset of SonVariantsChr21}
\description{
SonRareVariantsChr21 is a data frame consisting of annotated genomic variants
coming from CompleteGenomics whole genome sequencing. Variants were filtered from SonVariantsChr21 file using following filters: dbSNP frequency < 0.01, coding, nonsynonymous or nonsense variant, sequencing depth > 10

}
\usage{data(SonRareVariantsChr21)}
\value{test data}
\format{data frame}
\source{https://sites.stanford.edu/abms/content/giab-reference-materials-and-data}
\keyword{datasets}
